EW-644 activated exception filter for all controllers
