--- src/shared/error/schulconnex.error.ts
@@ -1,6 +1,0 @@
-export type SchulConnexError = {
-    statusCode: number;
-    subCode: string;
-    title: string;
-    description: string;
-};
